pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = ""
authors = ["Saveliy <s.lekhtin@innopolis.university>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
pydantic = {extras = ["email"], version = "^2.5.3"}
bcrypt = "^4.1.2"
python-multipart = "^0.0.6"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
pydantic-settings = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
